replace `~` with home directory
