+++ cmd/install.go
@@ -5,6 +5,7 @@ import (
 	"code-profiles/config"
 	"code-profiles/utils"
 
+	"github.com/marcantoineg/fileutil"
 	"github.com/spf13/cobra"
 )
 
@@ -12,7 +13,7 @@ var installCmd = &cobra.Command{
 	Use:   "install [profile_name]",
 	Short: "install required VSCode extensions for a given profile",
 	Run: func(cmd *cobra.Command, args []string) {
-		config.SetConfigPath(configPathFlag)
+		config.SetConfigPath(fileutil.ReplaceTilde(configPathFlag))
 
 		var profileName = ""
 		if len(args) >= 1 {
